pyproject.toml
[project]
name = "conspectius-engine"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "bitsandbytes>=0.49.2",
    "faster-whisper>=1.2.1",
    "langchain-text-splitters>=1.1.1",
    "loguru>=0.7.3",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "tenacity>=9.1.4",
    "torch>=2.5.1",
    "tqdm>=4.66.5",
    "transformers>=5.5.0",
]

[[tool.uv.index]]
url = "https://download.pytorch.org/whl/cu121"
